Fix 64bit

+++ WPFCaptureSample/WindowEnumerationHelper.cs
@@ -114,7 +114,7 @@ namespace WPFCaptureSample
 
         public static bool IsWindowValidForCapture(IntPtr hwnd)
         {
-            if (hwnd.ToInt32() == 0)
+            if (hwnd.ToInt64() == 0)
             {
                 return false;
             }
@@ -134,7 +134,7 @@ namespace WPFCaptureSample
                 return false;
             }
 
-            var style = (WindowStyles)(uint)GetWindowLongPtr(hwnd, (int)GWL.GWL_STYLE).ToInt32();
+            var style = (WindowStyles)(uint)GetWindowLongPtr(hwnd, (int)GWL.GWL_STYLE).ToInt64();
             if (style.HasFlag(WindowStyles.WS_DISABLED))
             {
                 return false;
